requirements.txt
Django==1.11.15
Django==2.1
Django==4.1.7
django_countries==7.5.1
django_filter==23.2
